requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0

